# backend/docs/openapi-governance.yaml
openapi: 3.1.0
info:
  title: DictionaryRoot Governed Platform API
  version: 0.2.0
  description: Authentication, account, governance, workflow, moderation, and audit additions to SourceRoot.
servers:
  - url: http://localhost:3000/api/v1
paths:
...
  /admin/overview:
    get:
      summary: Read governance operating metrics
      responses: { '200': { description: Overview }, '403': { description: Audit permission required } }
  /admin/organizations:
    get:
      summary: List organizations visible to the current administrator
      responses: { '200': { description: Scoped organization list } }
    post:
      summary: Create an organization; system permission is required
      responses: { '201': { description: Organization created }, '403': { description: System permission required } }
  /admin/organizations/{organizationId}/members:
    get:
      summary: List active members and organization-scoped roles for an administered organization
      parameters:
        - in: path
          name: organizationId
          required: true
          schema: { type: string, format: uuid }
      responses: { '200': { description: Member and scoped role list }, '403': { description: Organization permission required } }
  /admin/organizations/{organizationId}/members/{userId}/roles:
    post:
      summary: Assign an organization-scoped role to an active member
      parameters:
        - in: path
          name: organizationId
          required: true
          schema: { type: string, format: uuid }
        - in: path
          name: userId
          required: true
          schema: { type: string, format: uuid }
      responses: { '204': { description: Role assigned }, '403': { description: Organization permission required }, '409': { description: Invalid membership or role state } }
  /admin/organizations/{organizationId}/members/{userId}/roles/{roleKey}:
    delete:
      summary: Remove an organization-scoped role while preserving at least one organization administrator
      parameters:
        - in: path
          name: organizationId
          required: true
          schema: { type: string, format: uuid }
        - in: path
          name: userId
          required: true
          schema: { type: string, format: uuid }
        - in: path
          name: roleKey
          required: true
          schema: { type: string }
      responses: { '204': { description: Role removed }, '409': { description: Final organization administrator cannot be removed } }
  /admin/organizations/{organizationId}/invitations:
    post:
      summary: Invite a verified email to an administered organization with an organization-scoped role
      parameters:
        - in: path
          name: organizationId
          required: true
          schema: { type: string, format: uuid }
      responses: { '201': { description: Invitation created and delivered }, '403': { description: Organization permission required } }
  /admin/moderation/locks:
    get:
      summary: List active publication locks
      responses: { '200': { description: Active locks } }
    post:
      summary: Create a publication lock
      responses: { '201': { description: Lock created }, '409': { description: Target already locked } }
...
components:
  securitySchemes:
    dictionaryRootSession:
      type: apiKey
      in: cookie
      name: dr_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token